+++ packages/firestore/src/core/component_provider.ts
@@ -102,7 +102,8 @@ export interface OfflineComponentProvider {
  * Uses EagerGC garbage collection.
  */
 export class MemoryOfflineComponentProvider
-  implements OfflineComponentProvider {
+  implements OfflineComponentProvider
+{
   persistence!: Persistence;
   sharedClientState!: SharedClientState;
   localStore!: LocalStore;
@@ -195,8 +196,8 @@ export class IndexedDbOfflineComponentProvider extends MemoryOfflineComponentPro
   createGarbageCollectionScheduler(
     cfg: ComponentConfiguration
   ): GarbageCollectionScheduler | null {
-    const garbageCollector = this.persistence.referenceDelegate
-      .garbageCollector;
+    const garbageCollector =
+      this.persistence.referenceDelegate.garbageCollector;
     return new LruScheduler(garbageCollector, cfg.asyncQueue);
   }
 
@@ -261,10 +262,8 @@ export class MultiTabOfflineComponentProvider extends IndexedDbOfflineComponentP
           syncEngine
         ),
         getActiveClients: syncEngineGetActiveClients.bind(null, syncEngine),
-        synchronizeWithChangedDocuments: syncEngineSynchronizeWithChangedDocuments.bind(
-          null,
-          syncEngine
-        )
+        synchronizeWithChangedDocuments:
+          syncEngineSynchronizeWithChangedDocuments.bind(null, syncEngine)
       };
       await this.sharedClientState.start();
     }
@@ -347,10 +346,8 @@ export class OnlineComponentProvider {
         OnlineStateSource.SharedClientState
       );
 
-    this.remoteStore.remoteSyncer.handleCredentialChange = syncEngineHandleCredentialChange.bind(
-      null,
-      this.syncEngine
-    );
+    this.remoteStore.remoteSyncer.handleCredentialChange =
+      syncEngineHandleCredentialChange.bind(null, this.syncEngine);
 
     await remoteStoreApplyPrimaryState(
       this.remoteStore,