+++ packages/firestore/test/integration/util/internal_helpers.ts
@@ -129,18 +129,22 @@ function bundleWithTestDocsAndQueries(
   builder.addNamedQuery(
     'limit',
     { seconds: 1000, nanos: 9999 },
-    (collectionReference('coll-1')
-      .orderBy('bar', 'desc')
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      .limit(1) as any)._query
+    (
+      collectionReference('coll-1')
+        .orderBy('bar', 'desc')
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        .limit(1) as any
+    )._query
   );
   builder.addNamedQuery(
     'limit-to-last',
     { seconds: 1000, nanos: 9999 },
-    (collectionReference('coll-1')
-      .orderBy('bar', 'desc')
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      .limitToLast(1) as any)._query
+    (
+      collectionReference('coll-1')
+        .orderBy('bar', 'desc')
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        .limitToLast(1) as any
+    )._query
   );
 
   builder.addDocumentMetadata(a, { seconds: 1000, nanos: 9999 }, true);